+++ mail/mail/doctype/mail_queue/mail_queue.json
@@ -1,0 +1,374 @@
+{
+ "actions": [],
+ "creation": "2025-05-19 12:13:04.277760",
+ "doctype": "DocType",
+ "engine": "InnoDB",
+ "field_order": [
+  "section_break_zwhk",
+  "status",
+  "account",
+  "from_name",
+  "from_email",
+  "column_break_vzib",
+  "subject",
+  "section_break_agmx",
+  "save_as_draft",
+  "column_break_htqb",
+  "destroy_after_submission",
+  "reply_to_section",
+  "reply_to",
+  "headers_section",
+  "headers",
+  "recipients_section",
+  "recipients",
+  "body_section",
+  "html_body",
+  "text_body",
+  "attachments_section",
+  "attachments",
+  "more_info_tab",
+  "message_id",
+  "_id",
+  "blob_id",
+  "size",
+  "column_break_qxot",
+  "sent_at",
+  "column_break_kuap",
+  "in_reply_to",
+  "mailbox_id",
+  "thread_id",
+  "message_section",
+  "raw_message",
+  "message",
+  "response_section",
+  "_response",
+  "response",
+  "error_message_section",
+  "error_message"
+ ],
+ "fields": [
+  {
+   "fieldname": "section_break_zwhk",
+   "fieldtype": "Section Break"
+  },
+  {
+   "fieldname": "status",
+   "fieldtype": "Select",
+   "in_list_view": 1,
+   "in_standard_filter": 1,
+   "label": "Status",
+   "no_copy": 1,
+   "options": "\nDraft\nFailed to Draft\nSubmitted\nFailed to Submit\nFailed",
+   "read_only": 1,
+   "search_index": 1
+  },
+  {
+   "fieldname": "account",
+   "fieldtype": "Link",
+   "in_list_view": 1,
+   "in_standard_filter": 1,
+   "label": "Account",
+   "options": "Mail Account",
+   "reqd": 1,
+   "search_index": 1,
+   "set_only_once": 1
+  },
+  {
+   "fieldname": "from_name",
+   "fieldtype": "Data",
+   "in_list_view": 1,
+   "in_standard_filter": 1,
+   "label": "From Name",
+   "no_copy": 1,
+   "set_only_once": 1
+  },
+  {
+   "fieldname": "from_email",
+   "fieldtype": "Data",
+   "in_list_view": 1,
+   "in_standard_filter": 1,
+   "label": "From Email",
+   "no_copy": 1,
+   "search_index": 1,
+   "set_only_once": 1
+  },
+  {
+   "fieldname": "column_break_vzib",
+   "fieldtype": "Column Break"
+  },
+  {
+   "fieldname": "subject",
+   "fieldtype": "Small Text",
+   "in_standard_filter": 1,
+   "label": "Subject",
+   "set_only_once": 1
+  },
+  {
+   "collapsible": 1,
+   "collapsible_depends_on": "eval: doc.reply_to",
+   "fieldname": "reply_to_section",
+   "fieldtype": "Section Break",
+   "label": "Reply To"
+  },
+  {
+   "fieldname": "reply_to",
+   "fieldtype": "JSON",
+   "label": "Reply To",
+   "set_only_once": 1
+  },
+  {
+   "collapsible": 1,
+   "collapsible_depends_on": "eval: doc.headers",
+   "fieldname": "headers_section",
+   "fieldtype": "Section Break",
+   "label": "Headers"
+  },
+  {
+   "fieldname": "headers",
+   "fieldtype": "JSON",
+   "label": "Headers",
+   "set_only_once": 1
+  },
+  {
+   "fieldname": "recipients_section",
+   "fieldtype": "Section Break",
+   "label": "Recipients"
+  },
+  {
+   "fieldname": "recipients",
+   "fieldtype": "JSON",
+   "label": "Recipients",
+   "set_only_once": 1
+  },
+  {
+   "collapsible": 1,
+   "fieldname": "body_section",
+   "fieldtype": "Section Break",
+   "label": "Body"
+  },
+  {
+   "fieldname": "html_body",
+   "fieldtype": "Code",
+   "label": "HTML",
+   "set_only_once": 1
+  },
+  {
+   "fieldname": "text_body",
+   "fieldtype": "Code",
+   "label": "Text",
+   "set_only_once": 1
+  },
+  {
+   "collapsible": 1,
+   "collapsible_depends_on": "eval: doc.attachments",
+   "fieldname": "attachments_section",
+   "fieldtype": "Section Break",
+   "label": "Attachments"
+  },
+  {
+   "fieldname": "attachments",
+   "fieldtype": "JSON",
+   "label": "Attachments",
+   "set_only_once": 1
+  },
+  {
+   "fieldname": "message_id",
+   "fieldtype": "Data",
+   "in_standard_filter": 1,
+   "label": "Message ID",
+   "length": 255,
+   "no_copy": 1,
+   "search_index": 1,
+   "set_only_once": 1
+  },
+  {
+   "fieldname": "blob_id",
+   "fieldtype": "Data",
+   "label": "Blob ID",
+   "no_copy": 1,
+   "read_only": 1
+  },
+  {
+   "fieldname": "column_break_qxot",
+   "fieldtype": "Column Break"
+  },
+  {
+   "fieldname": "sent_at",
+   "fieldtype": "Datetime",
+   "label": "Sent At",
+   "no_copy": 1,
+   "search_index": 1,
+   "set_only_once": 1
+  },
+  {
+   "fieldname": "column_break_kuap",
+   "fieldtype": "Column Break"
+  },
+  {
+   "fieldname": "mailbox_id",
+   "fieldtype": "Data",
+   "in_standard_filter": 1,
+   "label": "Mailbox ID",
+   "no_copy": 1,
+   "read_only": 1,
+   "search_index": 1
+  },
+  {
+   "fieldname": "thread_id",
+   "fieldtype": "Data",
+   "in_standard_filter": 1,
+   "label": "Thread ID",
+   "no_copy": 1,
+   "read_only": 1,
+   "search_index": 1
+  },
+  {
+   "collapsible": 1,
+   "fieldname": "message_section",
+   "fieldtype": "Section Break",
+   "label": "Message"
+  },
+  {
+   "fieldname": "raw_message",
+   "fieldtype": "Code",
+   "label": "Raw Message",
+   "set_only_once": 1
+  },
+  {
+   "fieldname": "message",
+   "fieldtype": "Code",
+   "is_virtual": 1,
+   "label": "Message",
+   "no_copy": 1,
+   "read_only": 1
+  },
+  {
+   "fieldname": "response",
+   "fieldtype": "JSON",
+   "is_virtual": 1,
+   "label": "Response",
+   "no_copy": 1,
+   "read_only": 1
+  },
+  {
+   "fieldname": "_response",
+   "fieldtype": "JSON",
+   "hidden": 1,
+   "label": "Response",
+   "no_copy": 1,
+   "read_only": 1
+  },
+  {
+   "collapsible": 1,
+   "fieldname": "error_message_section",
+   "fieldtype": "Section Break",
+   "label": "Error Message"
+  },
+  {
+   "fieldname": "error_message",
+   "fieldtype": "Small Text",
+   "hidden": 1,
+   "is_virtual": 1,
+   "label": "Error Message",
+   "no_copy": 1,
+   "read_only": 1
+  },
+  {
+   "fieldname": "more_info_tab",
+   "fieldtype": "Tab Break",
+   "label": "More Info"
+  },
+  {
+   "collapsible": 1,
+   "fieldname": "response_section",
+   "fieldtype": "Section Break",
+   "label": "Response"
+  },
+  {
+   "default": "0",
+   "description": "When checked, the email will be saved as a draft instead of being sent.",
+   "fieldname": "save_as_draft",
+   "fieldtype": "Check",
+   "label": " Save as Draft",
+   "search_index": 1,
+   "set_only_once": 1
+  },
+  {
+   "fieldname": "section_break_agmx",
+   "fieldtype": "Section Break"
+  },
+  {
+   "fieldname": "column_break_htqb",
+   "fieldtype": "Column Break"
+  },
+  {
+   "default": "0",
+   "depends_on": "eval: !doc.save_as_draft",
+   "description": "If enabled, the message will be deleted after successful submission.",
+   "fieldname": "destroy_after_submission",
+   "fieldtype": "Check",
+   "label": "Destroy After Submission",
+   "set_only_once": 1
+  },
+  {
+   "fieldname": "in_reply_to",
+   "fieldtype": "Data",
+   "in_standard_filter": 1,
+   "label": "In Reply To (Message ID)",
+   "length": 255,
+   "search_index": 1,
+   "set_only_once": 1
+  },
+  {
+   "fieldname": "_id",
+   "fieldtype": "Data",
+   "in_standard_filter": 1,
+   "label": "Email ID",
+   "no_copy": 1,
+   "read_only": 1,
+   "search_index": 1
+  },
+  {
+   "fieldname": "size",
+   "fieldtype": "Int",
+   "label": "Size (Bytes)",
+   "no_copy": 1,
+   "non_negative": 1,
+   "read_only": 1
+  }
+ ],
+ "grid_page_length": 50,
+ "index_web_pages_for_search": 1,
+ "links": [],
+ "modified": "2025-05-19 15:50:30.326101",
+ "modified_by": "Administrator",
+ "module": "Mail",
+ "name": "Mail Queue",
+ "owner": "Administrator",
+ "permissions": [
+  {
+   "create": 1,
+   "delete": 1,
+   "email": 1,
+   "export": 1,
+   "print": 1,
+   "read": 1,
+   "report": 1,
+   "role": "System Manager",
+   "share": 1,
+   "write": 1
+  },
+  {
+   "create": 1,
+   "read": 1,
+   "role": "Mail User",
+   "write": 1
+  }
+ ],
+ "row_format": "Dynamic",
+ "show_title_field_in_link": 1,
+ "sort_field": "creation",
+ "sort_order": "DESC",
+ "states": [],
+ "title_field": "subject"
+}
